requirements.txt
pandas==1.1.5
numpy==1.19.5
pyEDFlib==0.1.22
click==8.0.1
py-ecg-detectors==1.0.2
wfdb==3.4.0
biosppy==0.7.3
hrv-analysis==1.0.4
#ecg-qc==1.0b4
great-expectations==0.13.25
airflow-provider-great-expectations==0.0.7
psycopg2-binary==2.8.6